requirements.txt
fastapi
uvicorn[standard]
playwright
aiohttp
python-dotenv
pandas
pdfplumber
beautifulsoup4
lxml
httpx
PyPDF2
